refactor: extract options flow mixin from config flow
