+++ custom_components/ecoflow_energy/config_flow.py
@@ -11,41 +11,19 @@ Step 5: Config entry created
 
 from __future__ import annotations
 
-import logging
 from typing import Any
 
-import aiohttp
-import voluptuous as vol
-
-from homeassistant.config_entries import ConfigEntry, ConfigFlow, ConfigFlowResult, OptionsFlow
+from homeassistant.config_entries import ConfigEntry, ConfigFlow, OptionsFlow
 from homeassistant.core import callback
-from homeassistant.helpers.aiohttp_client import async_get_clientsession
-from homeassistant.helpers.selector import SelectSelector, SelectSelectorConfig
 
-from .const import (
-    AUTH_METHOD_APP,
-    AUTH_METHOD_DEVELOPER,
-    CONF_ACCESS_KEY,
-    CONF_AUTH_METHOD,
-    CONF_DEVICES,
-    CONF_EMAIL,
-    CONF_MODE,
-    CONF_PASSWORD,
-    CONF_SECRET_KEY,
-    CONF_USER_ID,
-    DEVICE_TYPE_DISPLAY_NAMES,
-    DOMAIN,
-    MODE_ENHANCED,
-    MODE_STANDARD,
-    get_device_type,
-)
+from .config_flow_options import OptionsFlowMixin
 from .config_flow_reauth import ReauthFlowMixin
 from .config_flow_reconfigure import ReconfigureFlowMixin
-from .config_flow_setup import SetupFlowMixin, _device_label
-from .ecoflow.enhanced_auth import enhanced_login, get_app_device_list
-from .ecoflow.iot_api import IoTApiClient
+from .config_flow_setup import SetupFlowMixin
+from .const import AUTH_METHOD_DEVELOPER, DOMAIN, MODE_STANDARD
 
-_LOGGER = logging.getLogger(__name__)
+# hassfest requires config_flow to be a literal file - do not convert this
+# module into a config_flow/ package.
 
 
 class EcoFlowEnergyConfigFlow(
@@ -74,27 +52,7 @@ class EcoFlowEnergyConfigFlow(
         self._selected_devices: list[dict[str, Any]] = []
 
 
-async def _async_fetch_app_devices(
-    hass: Any, email: str, password: str
-) -> list[dict[str, Any]]:
-    """Log in to the app API and return the normalized device list.
-
-    Mirrors the discovery performed by
-    :meth:`EcoFlowEnergyConfigFlow.async_step_app_credentials`. Returns an
-    empty list when the login fails or the account exposes no devices.
-    Network and parsing errors are raised to the caller.
-    """
-    session = async_get_clientsession(hass)
-    login_result = await enhanced_login(session, email, password)
-    if login_result is None:
-        return []
-    raw_devices = await get_app_device_list(session, login_result["token"])
-    if not raw_devices:
-        return []
-    return EcoFlowEnergyConfigFlow._normalize_app_devices(raw_devices)
-
-
-class EcoFlowOptionsFlow(OptionsFlow):
+class EcoFlowOptionsFlow(OptionsFlowMixin, OptionsFlow):
     """Handle options for EcoFlow Energy."""
 
     def __init__(self) -> None:
@@ -102,247 +60,3 @@ class EcoFlowOptionsFlow(OptionsFlow):
         self._all_devices: list[dict[str, Any]] = []
         self._pending_mode: str = ""
         self._pending_devices: list[str] = []
-
-    async def async_step_init(
-        self, user_input: dict[str, Any] | None = None
-    ) -> ConfigFlowResult:
-        """Main options step — change mode and device selection."""
-        errors: dict[str, str] = {}
-
-        current_mode = self.config_entry.data.get(CONF_MODE, MODE_STANDARD)
-        current_device_sns = [
-            d["sn"] for d in self.config_entry.data.get(CONF_DEVICES, [])
-        ]
-
-        # Fetch current device list from API (developer auth only)
-        if not self._all_devices:
-            auth_method = self.config_entry.data.get(CONF_AUTH_METHOD, AUTH_METHOD_DEVELOPER)
-            if auth_method == AUTH_METHOD_DEVELOPER:
-                ak = self.config_entry.data.get(CONF_ACCESS_KEY)
-                sk = self.config_entry.data.get(CONF_SECRET_KEY)
-                if ak and sk:
-                    session = async_get_clientsession(self.hass)
-                    api = IoTApiClient(session, ak, sk)
-                    try:
-                        raw = await api.get_device_list()
-                        if raw:
-                            self._all_devices = EcoFlowEnergyConfigFlow._normalize_devices(raw)
-                    except (aiohttp.ClientError, TimeoutError, OSError, KeyError, ValueError, TypeError):
-                        _LOGGER.warning("Options flow: failed to fetch device list", exc_info=True)
-            elif auth_method == AUTH_METHOD_APP:
-                email = self.config_entry.data.get(CONF_EMAIL)
-                password = self.config_entry.data.get(CONF_PASSWORD)
-                if email and password:
-                    try:
-                        devices = await _async_fetch_app_devices(
-                            self.hass, email, password
-                        )
-                    except (
-                        aiohttp.ClientError,
-                        TimeoutError,
-                        OSError,
-                        KeyError,
-                        ValueError,
-                        TypeError,
-                        AttributeError,
-                    ):
-                        # Fall back to the stored device list - the options
-                        # flow stays usable without a fresh discovery.
-                        _LOGGER.debug(
-                            "Options flow: failed to fetch app device list",
-                            exc_info=True,
-                        )
-                    else:
-                        if devices:
-                            self._all_devices = devices
-                        else:
-                            _LOGGER.debug(
-                                "Options flow: app device list empty, "
-                                "using stored devices"
-                            )
-
-        if user_input is not None:
-            new_mode = user_input.get(CONF_MODE, current_mode)
-            selected_sns = user_input.get(CONF_DEVICES, current_device_sns)
-
-            if not selected_sns:
-                errors["base"] = "no_devices"
-            elif new_mode == MODE_ENHANCED and current_mode != MODE_ENHANCED:
-                # Switching to Enhanced - need email + password
-                self._pending_mode = new_mode
-                self._pending_devices = selected_sns
-                return await self.async_step_enhanced()
-            elif new_mode != MODE_ENHANCED and not self.config_entry.data.get(CONF_ACCESS_KEY):
-                # Switching to Standard but no Developer API keys stored
-                self._pending_mode = new_mode
-                self._pending_devices = selected_sns
-                return await self.async_step_developer()
-            else:
-                return self._save_options(new_mode, selected_sns)
-
-        if self._all_devices:
-            device_options = {
-                d["sn"]: _device_label(d)
-                for d in self._all_devices
-            }
-        else:
-            stored = {
-                d["sn"]: d.get("device_type", "")
-                for d in self.config_entry.data.get(CONF_DEVICES, [])
-            }
-            device_options = {
-                sn: f"{DEVICE_TYPE_DISPLAY_NAMES.get(stored.get(sn, ''), sn[:12])} ({sn[:12]})"
-                for sn in current_device_sns
-            }
-
-        return self.async_show_form(
-            step_id="init",
-            data_schema=vol.Schema(
-                {
-                    vol.Required(CONF_MODE, default=current_mode): vol.In(
-                        {
-                            MODE_STANDARD: "Standard - Official EcoFlow API",
-                            MODE_ENHANCED: "Enhanced - Real-time (~3 s)",
-                        }
-                    ),
-                    vol.Required(
-                        CONF_DEVICES,
-                        default=current_device_sns,
-                    ): SelectSelector(
-                        SelectSelectorConfig(
-                            options=[
-                                {"value": sn, "label": label}
-                                for sn, label in device_options.items()
-                            ],
-                            multiple=True,
-                        )
-                    ),
-                }
-            ),
-            errors=errors,
-        )
-
-    async def async_step_developer(
-        self, user_input: dict[str, Any] | None = None
-    ) -> ConfigFlowResult:
-        """Switch to Standard mode - collect Developer API keys."""
-        errors: dict[str, str] = {}
-
-        if user_input is not None:
-            access_key = user_input.get(CONF_ACCESS_KEY, "").strip()
-            secret_key = user_input.get(CONF_SECRET_KEY, "").strip()
-
-            if access_key and secret_key:
-                session = async_get_clientsession(self.hass)
-                api = IoTApiClient(session, access_key, secret_key)
-                try:
-                    devices = await api.get_device_list()
-                    if devices is not None:
-                        return self._save_options(
-                            self._pending_mode,
-                            self._pending_devices,
-                            access_key=access_key,
-                            secret_key=secret_key,
-                        )
-                    errors["base"] = "invalid_auth"
-                except (aiohttp.ClientError, TimeoutError, OSError):
-                    errors["base"] = "cannot_connect"
-            else:
-                errors["base"] = "invalid_auth"
-
-        return self.async_show_form(
-            step_id="developer",
-            data_schema=vol.Schema(
-                {
-                    vol.Required(CONF_ACCESS_KEY): str,
-                    vol.Required(CONF_SECRET_KEY): str,
-                }
-            ),
-            errors=errors,
-        )
-
-    async def async_step_enhanced(
-        self, user_input: dict[str, Any] | None = None
-    ) -> ConfigFlowResult:
-        """Switch to Enhanced mode - login to obtain userId."""
-        errors: dict[str, str] = {}
-
-        current_email = self.config_entry.data.get(CONF_EMAIL, "")
-
-        if user_input is not None:
-            email = user_input.get(CONF_EMAIL, "").strip()
-            password = user_input.get(CONF_PASSWORD, "")
-            user_id = ""
-
-            if email and password:
-                session = async_get_clientsession(self.hass)
-                try:
-                    login_result = await enhanced_login(session, email, password)
-                    if login_result is not None:
-                        user_id = login_result["user_id"]
-                except (aiohttp.ClientError, TimeoutError) as exc:
-                    _LOGGER.warning("Options flow: Enhanced login failed: %s", exc)
-                except (KeyError, ValueError, TypeError, AttributeError):
-                    _LOGGER.exception("Options flow: Enhanced login error")
-
-            if user_id:
-                return self._save_options(
-                    self._pending_mode,
-                    self._pending_devices,
-                    email=email,
-                    password=password,
-                    user_id=user_id,
-                )
-            else:
-                errors["base"] = "enhanced_login_failed"
-
-        return self.async_show_form(
-            step_id="enhanced",
-            data_schema=vol.Schema(
-                {
-                    vol.Required(CONF_EMAIL, default=current_email): str,
-                    vol.Required(CONF_PASSWORD): str,
-                }
-            ),
-            errors=errors,
-        )
-
-    def _save_options(
-        self,
-        mode: str,
-        selected_sns: list[str],
-        *,
-        email: str = "",
-        password: str = "",
-        user_id: str = "",
-        access_key: str = "",
-        secret_key: str = "",
-    ) -> ConfigFlowResult:
-        """Persist changes by updating config entry data."""
-        existing = {d["sn"]: d for d in self.config_entry.data.get(CONF_DEVICES, [])}
-        api = {d["sn"]: d for d in self._all_devices}
-
-        selected_devices = [api.get(sn) or existing.get(sn) for sn in selected_sns]
-        selected_devices = [d for d in selected_devices if d is not None]
-
-        new_data = dict(self.config_entry.data)
-        new_data[CONF_MODE] = mode
-        new_data[CONF_DEVICES] = selected_devices
-
-        if mode == MODE_ENHANCED:
-            new_data[CONF_AUTH_METHOD] = AUTH_METHOD_APP
-            new_data[CONF_EMAIL] = email or new_data.get(CONF_EMAIL, "")
-            new_data[CONF_PASSWORD] = password or new_data.get(CONF_PASSWORD, "")
-            new_data[CONF_USER_ID] = user_id or new_data.get(CONF_USER_ID, "")
-        else:
-            new_data[CONF_AUTH_METHOD] = AUTH_METHOD_DEVELOPER
-            if access_key:
-                new_data[CONF_ACCESS_KEY] = access_key
-            if secret_key:
-                new_data[CONF_SECRET_KEY] = secret_key
-            new_data.pop(CONF_EMAIL, None)
-            new_data.pop(CONF_PASSWORD, None)
-            new_data.pop(CONF_USER_ID, None)
-
-        self.hass.config_entries.async_update_entry(self.config_entry, data=new_data)
-        return self.async_create_entry(title="", data={})
